Removed common projects form hermes,bender into personal and removed need for AngryIP scanner

+++ dcjulian29.groups.personal/tools/chocolateyInstall.ps1
@@ -143,10 +143,11 @@ if (-not (Test-Path -Path "${env:TEMP}\dcjulian29.groups.personal.update.txt")) 
 
 # ~~~ Go language development stuff...
 
-go install github.com/goreleaser/goreleaser@latest
-go install github.com/spf13/cobra-cli@latest
-go install golang.org/x/lint/golint@latest
-
+if (-not (Test-Path -Path "${env:TEMP}\dcjulian29.groups.personal.update.txt")) {
+  go install github.com/goreleaser/goreleaser@latest
+  go install github.com/spf13/cobra-cli@latest
+  go install golang.org/x/lint/golint@latest
+}
 
 # ~~~ EOF
